UserService and UserServiceTest. Assertj in pom for test utils

+++ src/main/java/it/digiulio/social71/service/ICrudService.java
@@ -8,9 +8,9 @@ public interface ICrudService<T> {
 
     Set<T> findAll();
 
-    Optional<T> findById(int id);
+    Optional<T> findById(Long id);
 
     T update(T entity);
 
-    T delete(int id);
+    T delete(Long id);
 }
